+++ src/inputs/file/file.interfaces.ts
@@ -1,11 +1,12 @@
 import {BaseInputFieldDataInterface} from '../base/baseInput.interfaces'
 
 export interface FileInputFieldDataInterface extends BaseInputFieldDataInterface {
-	allowedFileTypes?: string[],
-	directUpload?: boolean,
-	hideFileName?: boolean,
-	maxFileSizeMB?: number,
-	preview?: boolean,
-	previewHeight?: string,
+	allowedFileTypes?: string[]
+	directUpload?: boolean
+	hideFileName?: boolean
+	maxFileSizeMB?: number
+	preview?: boolean
+	previewHeight?: string
+	previewIsRound?: boolean
 	previewWidth?: string
 }